refactor(core): move the WorkspaceReading seam into ClawCore
ClawAgent and ClawGateway imported the concrete ClawWorkspace module just to
reach the WorkspaceReading abstraction — the one violation of the seams-in-
ClawCore rule the module table draws. The protocol and the value types its
signatures need (WorkspaceFile, LoadedFile, SkillScanResult) now live in
ClawCore/Workspace/, ClawWorkspace keeps only FileSystemWorkspace, and the
two dependency edges are gone, so the package graph enforces the rule.

Also root-anchors .gitignore's workspace/ pattern: unanchored it matched
Sources/ClawCore/Workspace/ on a case-insensitive filesystem and silently
kept the new protocol file out of the repository.
