+++ Package.swift
@@ -66,7 +66,6 @@ let package = Package(
       name: "ClawAgent",
       dependencies: [
         "ClawCore",
-        "ClawWorkspace",
         .product(name: "Logging", package: "swift-log"),
       ]
     ),
@@ -76,7 +75,6 @@ let package = Package(
       dependencies: [
         "ClawCore",
         "ClawAgent",
-        "ClawWorkspace",
         .product(name: "ServiceLifecycle", package: "swift-service-lifecycle"),
         .product(name: "UnixSignals", package: "swift-service-lifecycle"),
         .product(name: "Logging", package: "swift-log"),
