+++ docs/ARCHITECTURE.md
@@ -61,7 +61,7 @@ State root: `~/.swift-claw/` (DB, secrets, logs) + `~/.swift-claw/workspace/` (i
 
 ## 3. Component architecture (modules)
 
-SwiftPM package `swift-claw`. Module prefix `Claw`; daemon binary `clawd`. The dependency graph is a **layered DAG**, not a star: stores are reached through **protocols declared in `ClawCore`**, so `ClawAgent` and `ClawGateway` depend only on `ClawCore` for persistence (concrete `ClawData` is injected at the composition root in `clawd`).
+SwiftPM package `swift-claw`. Module prefix `Claw`; daemon binary `clawd`. The dependency graph is a **layered DAG**, not a star: stores are reached through **protocols declared in `ClawCore`**, so `ClawAgent` and `ClawGateway` depend only on `ClawCore` for persistence (concrete `ClawData` is injected at the composition root in `clawd`). The rule generalizes: **seams live in `ClawCore`, implementations in sibling targets, composed at `clawd`** — e.g. `WorkspaceReading` is a `ClawCore` protocol; concrete `FileSystemWorkspace` (`ClawWorkspace`) is injected at the root.
 
 ```
 clawd
@@ -77,7 +77,7 @@ clawd
 
 | Target | Kind | Responsibility | Key types |
 |---|---|---|---|
-| `ClawCore` | lib | Pure domain: value types, config model, **error taxonomy** (§19), **store protocols**, tool/provider/transport/sandbox protocols, the shared **HTTP seam** (`HTTPExecuting`/`HTTPResult`, request + response headers — F16), and the 32 768-char `ReplySplitter`. No I/O. | `IncomingMessage`, `SessionKey`, `Principal`, `AppConfig`, `RunState`, `ApprovalState`, `RiskLevel`, `RunBudget`, `HTTPResult`, `ReplySplitter`, error taxonomy types; protocols: `LLMProvider`, `ChannelIntake`, `MessageDelivery`, `TelegramTransport` (composite), `HTTPExecuting`, `SecretStore`, `ExecutionBackend`, `Tool`, `SearchProviding`, `MessageStore`, `SessionStore`, `RunStore`, `AllowlistStore`, `UpdateCursorStore`, `OutboxStore`, `UsageStore`, `AuditLog`, `MemoryStore`, `ApprovalStore` |
+| `ClawCore` | lib | Pure domain: value types, config model, **error taxonomy** (§19), **store protocols**, tool/provider/transport/sandbox protocols, the shared **HTTP seam** (`HTTPExecuting`/`HTTPResult`, request + response headers — F16), and the 32 768-char `ReplySplitter`. No I/O. | `IncomingMessage`, `SessionKey`, `Principal`, `AppConfig`, `RunState`, `ApprovalState`, `RiskLevel`, `RunBudget`, `HTTPResult`, `ReplySplitter`, error taxonomy types; protocols: `LLMProvider`, `ChannelIntake`, `MessageDelivery`, `TelegramTransport` (composite), `HTTPExecuting`, `SecretStore`, `ExecutionBackend`, `Tool`, `SearchProviding`, `MessageStore`, `SessionStore`, `RunStore`, `AllowlistStore`, `UpdateCursorStore`, `OutboxStore`, `UsageStore`, `AuditLog`, `MemoryStore`, `ApprovalStore`, `WorkspaceReading` (+ its value types `WorkspaceFile`, `LoadedFile`, `SkillScanResult`) |
 | `ClawData` | lib | GRDB persistence: schema, `DatabaseMigrator`, store implementations. WAL. **Thin `Sendable` wrappers over `any DatabaseWriter`** (not actors guarding the pool); relies on GRDB serialization. | `Database`, concrete `…Store` types conforming to the `ClawCore` protocols |
 | `ClawTelegram` | lib | Thin Bot API client over AsyncHTTPClient; long-poll loop; envelope normalization; **`sendRichMessage` (markdown) + plain `sendMessage` fallback**; `sendChatAction`. (Escaping/splitter live in `ClawCore`.) | `TelegramClient`, `TelegramLongPoller`, `MessageEnvelope`, `InputRichMessage` |
 | `ClawLLM` | lib | OpenAI-compatible Chat Completions client + OpenAI-shaped message model; usage/cost; retries; **SSE streaming (v1)**. | `OpenAICompatibleProvider`, `ChatMessage`, `ChatRequest`, `ChatResponse`, `ToolCall`, `Usage`, `CostTable`, `SSEParser` |
